+++ tests/test_framework.c
@@ -92,7 +92,7 @@ double test_get_time_ms(void) {
 #else
     struct timeval tv;
     gettimeofday(&tv, NULL);
-    return tv.tv_sec * 1000.0 + tv.tv_usec / 1000.0;
+    return (double)tv.tv_sec * 1000.0 + (double)tv.tv_usec / 1000.0;
 #endif
 }
